+++ src/main/java/roomescape/controller/dto/request/ThemeCreateRequest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto;
+package roomescape.controller.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import org.hibernate.validator.constraints.URL;
