+++ src/main/java/roomescape/controller/dto/request/ThemeFamousFindRequest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto;
+package roomescape.controller.dto.request;
 
 import jakarta.validation.constraints.Positive;
 import java.time.LocalDate;
@@ -6,7 +6,7 @@ import java.time.LocalDate;
 public class ThemeFamousFindRequest {
     @Positive(message = "기간은 양수여야 합니다")
     private final Long days;
-    
+
     private final LocalDate date;
 
     @Positive(message = "개수는 양수여야 합니다")
