+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -6,8 +6,8 @@ import java.time.LocalDate;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.controller.dto.AvailableTimeFindRequest;
-import roomescape.controller.dto.ReservationTimeCreateRequest;
+import roomescape.controller.dto.request.AvailableTimeFindRequest;
+import roomescape.controller.dto.request.ReservationTimeCreateRequest;
 import roomescape.domain.reservation.ReservationTime;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
