+++ src/main/java/roomescape/service/ThemeService.java
@@ -6,8 +6,8 @@ import java.time.LocalDate;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.controller.dto.ThemeCreateRequest;
-import roomescape.controller.dto.ThemeFamousFindRequest;
+import roomescape.controller.dto.request.ThemeCreateRequest;
+import roomescape.controller.dto.request.ThemeFamousFindRequest;
 import roomescape.domain.theme.Theme;
 import roomescape.domain.theme.ThemeName;
 import roomescape.domain.theme.ThumbnailUrl;
