+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -15,8 +15,8 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import roomescape.controller.dto.ReservationCreateRequest;
-import roomescape.controller.dto.ReservationUpdateRequest;
+import roomescape.controller.dto.request.ReservationCreateRequest;
+import roomescape.controller.dto.request.ReservationUpdateRequest;
 import roomescape.domain.reservation.Reservation;
 import roomescape.domain.reservation.ReservationDate;
 import roomescape.domain.reservation.ReservationName;
