+++ src/test/java/roomescape/service/ReservationTimeServiceTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import roomescape.controller.dto.AvailableTimeFindRequest;
+import roomescape.controller.dto.request.AvailableTimeFindRequest;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
